requirements.txt
langchain
llama-index==0.10.14
llama-index-embeddings-huggingface==0.2.0  
fastapi
uvicorn
requests
python-dotenv
google-generativeai
psutil>=7.0.0
